update price to double, for different values; order products by price

+++ notyng/notyng/Feature/Dashboard/View/Cell/OrderTableViewCell.swift
@@ -189,7 +189,7 @@ class OrderTableViewCell: UITableViewCell {
     }
     
     private func setupTotalValue(products: [Product]) -> String {
-        var totalValue: Int = 0
+        var totalValue: Double = 0
         
         for product in products {
             totalValue = totalValue + product.price
